# docs/api.yml
openapi: 3.1.0
info:
  title: 文件服务 RESTful API
  version: 1.0.0
servers:
  - url: http://{host}/file/api/v1
    variables:
      host:
        default: localhost:8080
        description: 接入网关的域名与端口
paths:
  /files/{fkey}:
    get:
      summary: 下载文件
      parameters:
        - name: fkey
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 文件关联的外键
      responses:
        '200':
          description: 下载文件成功（文件元数据通过响应头返回，支持自定义头）
          headers:
            Content-Type:
              description: 实际文件的 MIME 类型（来自文件元数据）
              schema:
                type: string
                format: mime
            Content-Disposition:
              description: 建议的文件下载名称
              schema:
                type: string
            X-ICC-META:
              description: 文件元数据（JSON 字符串）。如需自定义元数据头，可使用 X-ICC-META-* 形式扩展。
              schema:
                type: string
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 删除文件
      parameters:
        - name: fkey
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 文件关联的外键
      responses:
        '204':
          description: 删除文件成功（无响应体）
        '403':
          description: 禁止删除文件
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                forbiddenDelete:
                  value:
                    code: 403
                    msg: 该文件禁止删除
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{fkey}/metadata:
    get:
      summary: 获取文件元数据
      parameters:
        - name: fkey
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 文件关联的外键
      responses:
        '200':
          description: 获取文件元数据成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMeta'
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{fkey}/presign:
    get:
      summary: 获取文件预签名下载 URL
      parameters:
        - name: fkey
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: 文件关联的外键
        - name: expiresIn
          in: query
          required: false
          schema:
            type: integer
            minimum: 60
            maximum: 86400
            default: 3600
          description: 预签名 URL 的有效期，单位为秒，默认值为 3600 秒(1 小时)
      responses:
        '200':
          description: 获取文件预签名下载 URL 成功
          content:
            text/plain:
              schema:
                type: string
                format: uri
                description: 预签名下载 URL
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          $ref: '#/components/responses/NotImplemented'
  /files/metadata:
    post:
      summary: 查询文件元数据
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从 0 开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 20
          description: 每页大小
        - name: sort
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
          description: 排序规则，格式为 属性,asc|desc，可重复传入多个
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 文件名
                name.prefix:
                  type: string
                  description: 文件名前缀匹配
                creator:
                  type: string
                  format: uuid, string
                  description: 文件创建者 ID或者 name
                creator.prefix:
                  type: string
                  description: 文件创建者名称前缀匹配
                tag.either:
                  type: string
                  description: 包含任意标签的文件，多个标签使用英文半角逗号 ',' 分隔
                tag.both:
                  type: string
                  description: 包含所有标签的文件，多个标签使用英文半角逗号 ',' 分隔
                created.at:
                  type: string
                  format: date
                  description: 文件创建时间等于指定时间的文件，格式为 YYYY-MM-DD
                created.before:
                  type: string
                  format: date
                  description: 文件创建时间早于指定时间的文件，格式为 YYYY-MM-DD
                created.after:
                  type: string
                  format: date
                  description: 文件创建时间晚于指定时间的文件，格式为 YYYY-MM-DD
                size:
                  type: integer
                  format: int64
                  description: 文件大小等于指定值的文件（字节）
                size.ge:
                  type: integer
                  format: int64
                  description: 文件大小大于等于指定值的文件（字节）
                size.gt:
                  type: integer
                  format: int64
                  description: 文件大小大于指定值的文件（字节）
                size.le:
                  type: integer
                  format: int64
                  description: 文件大小小于等于指定值的文件（字节）
                size.lt:
                  type: integer
                  format: int64
                  description: 文件大小小于指定值的文件（字节）
      responses:
        '200':
          description: 获取文件元数据成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFileMetaPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /files/upload:
    post:
      summary: 上传文件
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/FileMeta'
                - type: object
                  required:
                    - file
                    - createdBy
                    - fileName
                    - fileType
                  properties:
                    file:
                      type: string
                      format: binary
                      description: 要上传的文件
                    fileSize:
                      type: integer
                      format: int64
                      description: 文件大小（字节）。默认值 -1 表示上传完成后使用实际接收的字节流大小（可能因传输不稳定而不准确）。
                      default: -1
      responses:
        '200':
          description: 文件上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFileUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务或 HTTP 状态码
          example: 200
        msg:
          type: string
          description: 响应消息
          example: 成功
        data:
          description: 业务响应数据
      required:
        - code
        - msg
    ApiResponseFileUpload:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/FileUploadResult'
          required:
            - data
...
    ApiResponseFileMetaPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PageFileMeta'
          required:
            - data
...
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 业务或 HTTP 错误码
          example: 400
        msg:
          type: string
          description: 错误信息
          example: 请求参数错误
        data:
          type: object
          description: 业务响应数据（错误响应通常为空）
      required:
        - code
        - msg
    FileMeta:
      type: object
      properties:
        createdBy:
          type: string
          format: uuid
          description: 文件创建者 ID
        creatorName:
          type: string
          description: 文件创建者名称
        tags:
          type: string
          description: "文件标签列表，使用英文半角逗号 ',' 分隔"
        eTag:
          type: string
          description: 文件的 eTag
        fileName:
          type: string
          description: 文件名称
        fileType:
          type: string
          format: mime
          description: 文件类型(MIME 类型)
          example: image/png
        fkey:
          type: string
          format: uuid
          description: 文件关联的外键
        public:
          type: boolean
          description: 文件是否公开, 公开的文件可以使用形如http://{host}/static/api/v1/files/{fkey}的链接访问
        customMetadata:
          type: object
          description: 文件的自定义元数据（键值均为字符串）
          additionalProperties:
            type: string
    PageFileMeta:
      type: object
      description: Spring Data Page<FileMeta>
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FileMeta'
        totalElements:
          type: integer
          format: int64
          description: 总元素数
        totalPages:
          type: integer
          description: 总页数
        size:
          type: integer
          description: 每页大小
        number:
          type: integer
          description: 当前页码（从 0 开始）
        numberOfElements:
          type: integer
          description: 当前页元素数量
        first:
          type: boolean
          description: 是否为首页
        last:
          type: boolean
          description: 是否为末页
        empty:
          type: boolean
          description: 是否为空页
        sort:
          type: object
          description: 排序信息
    FileUploadResult:
      type: object
      properties:
        fkey:
          type: string
          format: uuid
          description: 文件关联的外键
        createdAt:
          type: string
          format: iso8601
          description: 文件创建时间
          example: "2024-01-01T12:00:00Z"
        fileUrl:
          type: string
          format: uri
          description: 已上传文件的访问 URL（基于接入网关的 host）
          example: "http://{host}/file/api/v1/files/123e4567-e89b-12d3-a456-426614174000"
      required:
        - fkey
        - createdAt
        - fileUrl
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            invalidUpload:
              value:
                code: 400
                msg: 文件上传信息无效
    NotFound:
      description: 文件未找到
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            fileNotFound:
              value:
                code: 404
                msg: 文件未找到
            taskNotFound:
              value:
                code: 404
                msg: 上传任务未找到
    NotImplemented:
      description: 功能未实现（当前存储类型不支持）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            notImplemented:
              value:
                code: 501
                msg: 当前存储类型不支持该功能
    PayloadTooLarge:
      description: 请求体过大
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            fileTooLarge:
              value:
                code: 413
                msg: 文件大小超过最大限制
            metaDataTooLarge:
              value:
                code: 413
                msg: 自定义元数据大小超过最大限制
    TooManyRequests:
      description: 请求过于频繁
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            rateLimited:
              value:
                code: 429
                msg: 触发限流，请稍后再试
    InternalServerError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            unexpected:
              value:
                code: 500
                msg: 处理文件上传时发生未知错误
    ServiceUnavailable:
      description: 服务不可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            unavailable:
              value:
                code: 503
                msg: 服务暂不可用，请稍后再试
    GatewayTimeout:
      description: 网关超时
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            timeout:
              value:
                code: 504
                msg: 服务器响应超时，请稍后再试